adapted function setLineAttributes() in ivtobjectlineattributes_c.h new line colour was passed as a uint16_t but in vtobjectlineattributes_c.h passed as uint8_t

+++ library/xgpl_src/IsoAgLib/comm/Part6_VirtualTerminal_Client/ivtobjectlineattributes_c.h
@@ -107,11 +107,11 @@ public:
   ///
   /// @param newLineColour
   /// @param b_updateObject
-  void setLineAttributes(uint16_t newLineColour, uint8_t newLineWidth, uint16_t newLineArt, bool b_updateObject=false, bool b_enableReplaceOfCmd=false) {
+  void setLineAttributes(uint8_t newLineColour, uint8_t newLineWidth, uint16_t newLineArt, bool b_updateObject=false, bool b_enableReplaceOfCmd=false) {
     vtObjectLineAttributes_c::setLineAttributes (newLineColour, newLineWidth, newLineArt, b_updateObject, b_enableReplaceOfCmd);
   }
 
-#ifdef USE_ISO_TERMINAL_GETATTRIBUTES  
+#ifdef USE_ISO_TERMINAL_GETATTRIBUTES
   /** that attribute is in parentheses in the spec, so commented out here
   uint8_t updateObjectType() const { return vtObjectLineAttributes_c::updateObjectType(); }
   */
